# lib/api-zod/src/generated/types/membershipPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PropFlow - Property Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MembershipPlan {
  id: number;
  name: string;
  slug: string;
  description: string;
  priceMonthly: number;
  /** @nullable */
  priceYearly?: number | null;
  features: string[];
  /** @nullable */
  maxSearches?: number | null;
  /** @nullable */
  maxSavedProperties?: number | null;
  hasForeclosures: boolean;
  hasTaxDelinquent: boolean;
  hasCodeViolations: boolean;
  isPopular: boolean;
}