# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PropFlow - Property Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyStatus } from "./propertyStatus";
import type { PropertyType } from "./propertyType";

export interface Property {
  id: number;
  type: PropertyType;
  status: PropertyStatus;
  address: string;
  city: string;
  state: string;
  zip: string;
  /** @nullable */
  county?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  estimatedValue?: number | null;
  /** @nullable */
  beds?: number | null;
  /** @nullable */
  baths?: number | null;
  /** @nullable */
  sqft?: number | null;
  /** @nullable */
  yearBuilt?: number | null;
  /** @nullable */
  lotSize?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  auctionDate?: string | null;
  /** @nullable */
  taxAmount?: number | null;
  /** @nullable */
  violationCode?: string | null;
  /** @nullable */
  violationDescription?: string | null;
  /** @nullable */
  postedByUserId?: string | null;
  createdAt: string;
  updatedAt: string;
}