# lib/api-zod/src/generated/types/subscribePlanBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PropFlow - Property Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SubscribePlanBodyBillingCycle } from "./subscribePlanBodyBillingCycle";

export interface SubscribePlanBody {
  planId: number;
  billingCycle: SubscribePlanBodyBillingCycle;
}